requirements.txt
torchvision
torch 
hydra-core
omegaconf
pytorch_lightning